backend/requirements.txt
fastapi
uvicorn[standard]
requests
python-dotenv
pandas
python-multipart